requirements.txt
# 애니메이션 메타데이터 자동 입력 시스템 의존성
# Python 3.8+ 필요

# 핵심 라이브러리
openai>=1.0.0                    # OpenAI API (Assistant, GPT-4.1)
requests>=2.31.0                 # HTTP 요청 (Notion API)
python-dotenv>=1.0.0             # 환경변수 관리

# 라프텔 API (비공식)
laftel>=1.3.0                    # 라프텔 비공식 API 래퍼
aiohttp>=3.8.0                   # laftel 의존성

# 표준 라이브러리 (별도 설치 불필요)
# json, os, re, csv, glob, argparse, subprocess
# datetime, typing, sys
